% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's algorithm for connecting a source to one or more targets}
\usage{
dijkstra(src, trgts, fscore = lpgm, max_nodes = 100, min_ev = 3)
}
\arguments{
\item{src}{Source gene}

\item{trgts}{Vector of target genes}

\item{fscore}{Scoring function that accepts a vector of evidence counts}

\item{max_nodes}{Maximum number of nodes to expand (Default: 100)}

\item{min_ev}{Minimum evidence count to consider}
}
\value{
A set of paths from src to each element of trgts
}
\description{
Runs Dijkstra's graph search algorithm using the provided path scoring function
}
