% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indra.R
\name{IDBquery}
\alias{IDBquery}
\title{INDRA DB REST query}
\usage{
IDBquery(...)
}
\arguments{
\item{...}{parameters to pass to indra.sources.indra_db_rest.get_statements()}
}
\value{
Data frame of edge properties
}
\description{
Performs an INDRA DB REST query with best_first=TRUE, ev_limit=1
Passes the result through queryEdges()
}
