% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{lpgm}
\alias{lpgm}
\title{Length-penalized geometric mean}
\usage{
lpgm(v)
}
\arguments{
\item{v}{Vector of edge weights}
}
\value{
Path score
}
\description{
Computes geometric mean of edge weights and offsets it by path length
}
