% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotPaths}
\alias{plotPaths}
\title{Plot network paths}
\usage{
plotPaths(...)
}
\arguments{
\item{...}{One or more Path data frames, as returned by, e.g., dijkstra()}
}
\value{
A ggplot2 object for additional manipulation
}
\description{
Plotting of paths discovered by graph search algorithms
}
