% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indra.R
\name{queryEdges}
\alias{queryEdges}
\title{Characterize edges retrieved by an INDRA DB REST query}
\usage{
queryEdges(indra_query)
}
\arguments{
\item{indra_query}{Query object returned by indra.source.indra_db_rest.get_statements()}
}
\value{
Data frame of edge properties
}
\description{
Given an INDRA DB REST query object, retrieves the set of Source-Target edges
}
