% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{spgm}
\alias{spgm}
\title{Strongly-penalized geometric mean}
\usage{
spgm(v)
}
\arguments{
\item{v}{Vector of edge weights}
}
\value{
Path score
}
\description{
Geometric mean of edge weights with a strong penalty for path length
}
